Generalised rules for comparables:
- Removed all obsolete boilerplate classes as a more generic solution is put in place.
- Corrected relevant existing rule by flattening the ineheritance structure
