--- BalancedFieldLength/Core/src/Core.Common.Data/DataModel/ValidationRules/DoubleParameterGreaterOrEqualToRule.cs
@@ -1,56 +1,0 @@
-﻿// Copyright (C) 2018 Dennis Tang. All rights reserved.
-//
-// This file is part of Balanced Field Length.
-//
-// Balanced Field Length is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-// 
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with this program. If not, see <http://www.gnu.org/licenses/>.
-
-using Core.Common.Data.Properties;
-
-namespace Core.Common.Data.DataModel.ValidationRules
-{
-    /// <summary>
-    /// Validation rule representing a parameter greater or equal to a lower limit.
-    /// </summary>
-    public class DoubleParameterGreaterOrEqualToRule : DoubleParameterRuleBase
-    {
-        private readonly double lowerLimit;
-
-        /// <inheritdoc/>
-        /// <summary>
-        /// Creates a new instance of <see cref="DoubleParameterGreaterOrEqualToRule"/>.
-        /// </summary>
-        /// <param name="lowerLimit">The lower limit value which the <paramref name="value"/>
-        /// must be greater or equal to.</param>
-        public DoubleParameterGreaterOrEqualToRule(string parameterName, double value, double lowerLimit)
-            : base(parameterName, value)
-        {
-            this.lowerLimit = lowerLimit;
-        }
-
-        public override ValidationRuleResult Execute()
-        {
-            ValidationRuleResult result = ValidateValueConcreteNumber();
-            if (result != ValidationRuleResult.ValidResult)
-            {
-                return result;
-            }
-
-            return !(Value >= lowerLimit) 
-                       ? ValidationRuleResult.CreateInvalidResult(
-                           string.Format(Resources.NumericRule_Value_0_must_be_greater_or_equal_to_LowerLimit_1_Current_value_2, 
-                                         ParameterName, lowerLimit, Value)) 
-                       : ValidationRuleResult.ValidResult;
-        }
-    }
-}
