+++ BalancedFieldLength/Core/test/Core.Common.Data.Test/DataModel/ValidationRules/DoubleParameterConcreteNumberRuleTest.cs
@@ -23,8 +23,25 @@ using NUnit.Framework;
 namespace Core.Common.Data.Test.DataModel.ValidationRules
 {
     [TestFixture]
-    public class DoubleParameterConcreteNumberRuleTest : DoubleParameterRuleBaseTestFixture<DoubleParameterConcreteNumberRule>
+    public class DoubleParameterConcreteNumberRuleTest
     {
+        [Test]
+        public void Constructor_WithArguments_ExpectedValues()
+        {
+            // Setup
+            const string parameterName = "ParameterName";
+
+            var random = new Random(21);
+            double lowerLimit = random.NextDouble();
+            double value = random.NextDouble();
+
+            // Call
+            var rule = new DoubleParameterConcreteNumberRule(parameterName, value);
+
+            // Assert
+            Assert.That(rule, Is.InstanceOf<ParameterRuleBase>());
+        }
+
         [Test]
         public void Execute_WithValidValue_ReturnsValidResult()
         {
@@ -43,9 +60,24 @@ namespace Core.Common.Data.Test.DataModel.ValidationRules
             Assert.That(result, Is.SameAs(ValidationRuleResult.ValidResult));
         }
 
-        protected override DoubleParameterConcreteNumberRule CreateRule(string parameterName, double value)
+        [Test]
+        [TestCase(double.NaN)]
+        [TestCase(double.NegativeInfinity)]
+        [TestCase(double.PositiveInfinity)]
+        public void Execute_WithNonConcreteValues_ReturnsExpectedValidationResult(double invalidValue)
         {
-            return new DoubleParameterConcreteNumberRule(parameterName, value);
+            // Setup
+            const string parameterName = "ParameterName";
+            var rule = new DoubleParameterConcreteNumberRule(parameterName, invalidValue);
+
+            // Call 
+            ValidationRuleResult result = rule.Execute();
+
+            // Assert
+            Assert.That(result.IsValid, Is.False);
+
+            var expectedMessage = $"{parameterName} must be a concrete number.";
+            Assert.That(result.ValidationMessage, Is.EqualTo(expectedMessage));
         }
     }
 }
